add class quickSort
